fix some bugs, add verification bash on yuvdns

+++ lut_cases_lsc.sh
@@ -5,6 +5,7 @@ copy_count=0
 modify_vlut_count=0
 modify_clut_count=0
 
+
 # Process each folder
 for folder in */; do
     # Get folder name
@@ -12,7 +13,7 @@ for folder in */; do
     
     # Navigate to the config folder
     config_folder="$folder/Cmodel/run/config"
-    if [ -d "$config_folder" ]; then
+    if [ -d "$config_folder" ] && [[ "$folder_name" != *bypass ]]; then
         # Delete specific file
         if [ -f "$config_folder/lsc858_12fracbit_tb.txt" ]; then
             rm "$config_folder/lsc858_12fracbit_tb.txt"
@@ -31,14 +32,14 @@ for folder in */; do
                 cp "./auto_lsc_config/lsc651_12fracbit_tb.txt" "$config_folder/"
                 echo "Copied lsc651_12fracbit_tb.txt to $config_folder for folder $folder_name"
                 ((copy_count++))
-                sed -i '31s|lsc858_12fracbit_tb|lsc651_12fracbit_tb|' "$folder/vmodel/sim/tb_test.v"
+                sed -i 's|lsc858_12fracbit_tb|lsc651_12fracbit_tb|g' "$folder/vmodel/sim/tb_test.v"
                 if [ $? -eq 0 ]; then
                     echo "Modified tb_test.v in $folder_name for lut2"
                     ((modify_vlut_count++))
                 else
                     echo "Failed to modify tb_test.v in $folder_name for lut2"
                 fi
-                sed -i '58,59s|lsc858_12fracbit_tb.txt|lsc651_12fracbit_tb.txt|' "$folder_name/Cmodel/run/xkisp_HDR.sh"
+                sed -i 's|lsc858_12fracbit_tb.txt|lsc651_12fracbit_tb.txt|g' "$folder/Cmodel/run/xkisp_HDR.sh"
                 if [ $? -eq 0 ]; then
                     echo "Modified xkisp_HDR.sh in $folder_name for lut2"
                     ((modify_clut_count++))
@@ -51,14 +52,14 @@ for folder in */; do
                 cp "./auto_lsc_config/lsc383_12fracbit_tb.txt" "$config_folder/"
                 echo "Copied lsc383_12fracbit_tb.txt to $config_folder for folder $folder_name"
                 ((copy_count++))
-                sed -i '31s|lsc858_12fracbit_tb|lsc383_12fracbit_tb|' "$folder/vmodel/sim/tb_test.v"
+                sed -i 's|lsc858_12fracbit_tb|lsc383_12fracbit_tb|g' "$folder/vmodel/sim/tb_test.v"
                 if [ $? -eq 0 ]; then
                     echo "Modified tb_test.v in $folder_name for lut3"
                     ((modify_vlut_count++))
                 else
                     echo "Failed to modify tb_test.v in $folder_name for lut3"
                 fi
-                sed -i '58,59s|lsc858_12fracbit_tb.txt|lsc383_12fracbit_tb.txt|' "$folder_name/Cmodel/run/xkisp_HDR.sh"
+                sed -i 's|lsc858_12fracbit_tb.txt|lsc383_12fracbit_tb.txt|g' "$folder/Cmodel/run/xkisp_HDR.sh"
                 if [ $? -eq 0 ]; then
                     echo "Modified xkisp_HDR.sh in $folder_name for lut3"
                     ((modify_clut_count++))
@@ -72,14 +73,14 @@ for folder in */; do
                 cp "./auto_lsc_config/lscmax_tb.txt" "$config_folder/"
                 echo "Copied lscmax_tb.txt to $config_folder for folder $folder_name"
                 ((copy_count++))
-                sed -i '31s|lsc858_12fracbit_tb|lscmax_tb|' "$folder/vmodel/sim/tb_test.v"
+                sed -i 's|lsc858_12fracbit_tb|lscmax_tb|g' "$folder/vmodel/sim/tb_test.v"
                 if [ $? -eq 0 ]; then
                     echo "Modified tb_test.v in $folder_name for lutmax"
                     ((modify_vlut_count++))
                 else
                     echo "Failed to modify tb_test.v in $folder_name for lutmax"
                 fi
-                sed -i '58,59s|lsc858_12fracbit_tb.txt|lscmax_tb.txt|' "$folder_name/Cmodel/run/xkisp_HDR.sh"
+                sed -i 's|lsc858_12fracbit_tb.txt|lscmax_tb.txt|g' "$folder/Cmodel/run/xkisp_HDR.sh"
                 if [ $? -eq 0 ]; then
                     echo "Modified xkisp_HDR.sh in $folder_name for lutmax"
                     ((modify_clut_count++))
@@ -92,14 +93,14 @@ for folder in */; do
                 cp "./auto_lsc_config/lscmin_tb.txt" "$config_folder/"
                 echo "Copied lscmin_tb.txt to $config_folder for folder $folder_name"
                 ((copy_count++))
-                sed -i '31s|lsc858_12fracbit_tb|lscmin_tb|' "$folder/vmodel/sim/tb_test.v"
+                sed -i 's|lsc858_12fracbit_tb|lscmin_tb|g' "$folder/vmodel/sim/tb_test.v"
                 if [ $? -eq 0 ]; then
                     echo "Modified tb_test.v in $folder_name for lutmin"
                     ((modify_vlut_count++))
                 else
                     echo "Failed to modify tb_test.v in $folder_name for lutmin"
                 fi
-                sed -i '58,59s|lsc858_12fracbit_tb.txt|lscmin_tb.txt|' "$folder_name/Cmodel/run/xkisp_HDR.sh"
+                sed -i 's|lsc858_12fracbit_tb.txt|lscmin_tb.txt|g' "$folder/Cmodel/run/xkisp_HDR.sh"
                 if [ $? -eq 0 ]; then
                     echo "Modified xkisp_HDR.sh in $folder_name for lutmin"
                     ((modify_clut_count++))
@@ -112,14 +113,14 @@ for folder in */; do
                 cp "./auto_lsc_config/lscran_tb.txt" "$config_folder/"
                 echo "Copied lscran_tb.txt to $config_folder for folder $folder_name"
                 ((copy_count++))
-                sed -i '31s|lsc858_12fracbit_tb|lscran_tb|' "$folder/vmodel/sim/tb_test.v"
+                sed -i 's|lsc858_12fracbit_tb|lscran_tb|g' "$folder/vmodel/sim/tb_test.v"
                 if [ $? -eq 0 ]; then
                     echo "Modified tb_test.v in $folder_name for lutran"
                     ((modify_vlut_count++))
                 else
                     echo "Failed to modify tb_test.v in $folder_name for lutran"
                 fi
-                sed -i '58,59s|lsc858_12fracbit_tb.txt|lscran_tb.txt|' "$folder_name/Cmodel/run/xkisp_HDR.sh"
+                sed -i 's|lsc858_12fracbit_tb.txt|lscran_tb.txt|g' "$folder/Cmodel/run/xkisp_HDR.sh"
                 if [ $? -eq 0 ]; then
                     echo "Modified xkisp_HDR.sh in $folder_name for lutran"
                     ((modify_clut_count++))
